Entities adapted to real database, and codelize service implemented to initiate scan from a given analysis id

+++ src/main/java/mx/com/solucionestea/codelizer/database/dao/AnalysisDao.java
@@ -10,5 +10,5 @@ import javax.transaction.Transactional;
  * Created by giovanni on 7/12/16.
  */
 @Transactional
-public interface AnalysisDao extends CrudRepository<Analysis, Long> {
+public interface AnalysisDao extends CrudRepository<Analysis, Integer> {
 }
